backend/requirements.txt
spotipy
pandas
scikit-learn
streamlit
python-dotenv